# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Underworld game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  username: string;
  rank: string;
  rankProgress: number;
  hp: number;
  maxHp: number;
  energy: number;
  maxEnergy: number;
  nerve: number;
  maxNerve: number;
  cash: number;
  respect: number;
  streakDays: number;
  city: string;
}

export interface DailyReward {
  canClaim: boolean;
  nextClaimInSeconds: number;
  streakDays: number;
  rewardAmount: number;
}

export interface DailyRewardResult {
  success: boolean;
  amount: number;
  newStreakDays: number;
}

export interface Crime {
  id: string;
  name: string;
  description: string;
  cashReward: number;
  respReward: number;
  nerveCost: number;
  successRate: number;
}

export interface CrimeResult {
  success: boolean;
  message: string;
  cashGained: number;
  respectGained: number;
}

export interface CarTarget {
  id: string;
  make: string;
  model: string;
  value: number;
  difficultyPct: number;
}

export interface CarTheftResult {
  success: boolean;
  message: string;
  /** @nullable */
  vehicle?: string | null;
  cashValue: number;
}

export interface HeistPlan {
  id: string;
  name: string;
  description: string;
  reward: number;
  nerveCost: number;
  minPlayers: number;
}

export interface HeistInput {
  planId: string;
}

export interface HeistResult {
  success: boolean;
  message: string;
  cashGained: number;
  respectGained: number;
}

export interface City {
  id: string;
  name: string;
  country: string;
  travelCost: number;
  unlockRank: string;
}

export interface TravelInput {
  cityId: string;
}

export interface TravelResult {
  success: boolean;
  message: string;
  newCity: string;
}

export interface ActivityEntry {
  id: number;
  username: string;
  action: string;
  detail: string;
  timeAgo: string;
  avatarColor: string;
}

export interface RecentActivityEntry {
  id: number;
  description: string;
  /** @nullable */
  target?: string | null;
  amount: number;
  timeAgo: string;
  type: string;
}

export interface Vehicle {
  id: number;
  make: string;
  model: string;
  year: number;
  value: number;
  condition: string;
}

export interface RankingEntry {
  position: number;
  username: string;
  rank: string;
  respect: number;
  cash: number;
}

export interface Property {
  id: string;
  name: string;
  description: string;
  cost: number;
  incomePerHour: number;
  category: string;
  owned: boolean;
  pendingIncome: number;
  /** @nullable */
  lastCollectedAt: string | null;
}

export interface BuyPropertyResult {
  success: boolean;
  message: string;
}

export interface CollectPropertyResult {
  success: boolean;
  amountCollected: number;
}

export interface MarketItem {
  id: string;
  name: string;
  description: string;
  price: number;
  type: string;
  effect: string;
  owned: boolean;
}

export interface BuyItemResult {
  success: boolean;
  message: string;
}

export interface Message {
  id: number;
  senderUsername: string;
  content: string;
  read: boolean;
  timeAgo: string;
}

export interface SendMessageInput {
  toUsername: string;
  content: string;
}

export interface SendMessageResult {
  success: boolean;
}

export interface MarkMessageReadResult {
  success: boolean;
}

export interface GroupCrime {
  id: string;
  name: string;
  description: string;
  cashReward: number;
  nerveCost: number;
  minMembers: number;
  successRate: number;
}

export interface CommitGroupCrimeResult {
  success: boolean;
  message: string;
  cashGained: number;
  respectGained: number;
}