# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Underworld game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties/{propertyId}/buy:
    post:
      operationId: buyProperty
      tags: [properties]
      summary: Purchase a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Purchase result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuyPropertyResult"
  /properties/{propertyId}/collect:
    post:
      operationId: collectPropertyIncome
      tags: [properties]
      summary: Collect pending income from a property
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectPropertyResult"
...
  /market/buy/{itemId}:
    post:
      operationId: buyMarketItem
      tags: [market]
      summary: Buy a black market item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Purchase result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuyItemResult"
...
  /messages/send:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to another player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResult"
  /messages/{messageId}/read:
    post:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark a message as read
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkMessageReadResult"
...
  /group-crimes/{crimeId}/commit:
    post:
      operationId: commitGroupCrime
      tags: [groupCrimes]
      summary: Commit a group crime
      parameters:
        - name: crimeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitGroupCrimeResult"
components:
  schemas:
...
    BuyPropertyResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    CollectPropertyResult:
      type: object
      required: [success, amountCollected]
      properties:
        success:
          type: boolean
        amountCollected:
          type: integer
...
    BuyItemResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    SendMessageInput:
      type: object
      required: [toUsername, content]
      properties:
        toUsername:
          type: string
        content:
          type: string
    SendMessageResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    MarkMessageReadResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    CommitGroupCrimeResult:
      type: object
      required: [success, message, cashGained, respectGained]
      properties:
        success:
          type: boolean
        message:
          type: string
        cashGained:
          type: integer
        respectGained:
          type: integer