# lib/api-zod/src/generated/types/buyItemResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Underworld game API
 * OpenAPI spec version: 0.1.0
 */

export interface BuyItemResult {
  success: boolean;
  message: string;
}