# lib/api-zod/src/generated/types/commitGroupCrimeResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Underworld game API
 * OpenAPI spec version: 0.1.0
 */

export interface CommitGroupCrimeResult {
  success: boolean;
  message: string;
  cashGained: number;
  respectGained: number;
}